Update MCommand.h
re-added sanity checks on mcommandparameterstring/blob
Quick fix for memory leak caused by MakeParam template skipping CMemPool::operator new, combined with the destructor inheritance on MCommandParameters - CMemPool::operator delete is called instead of deallocate or ::delete.
Therefore attempting to un-link memory that was never linked to the mempool in the first place.

Bottom line, MakeParam allocation and construction + the way MCommandParameters are destroyed = not re-using the memory, and not freeing it either